refactor(kinds): read the resource-kind tables from lib, not the browser
Completes the move: every importer now takes these from lib/kinds, so
ResourceBrowser stops re-exporting data it happens to render.

App's test mocked ResourceBrowser to supply a reduced set of five kinds. That
mock moves to lib/kinds with the tables, otherwise the sidebar would have
rendered the real forty and the assertions would have been testing a different
thing than they were written for.
